services/auth/requirements.txt
bcrypt==4.1.3
cryptography==42.0.5
fastapi==0.110.2
passlib==1.7.4
pika==1.3.2
pydantic==2.7.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
uvicorn==0.29.0
annotated-types==0.6.0
dnspython==2.6.1
email_validator==2.1.1
idna==3.7
pydantic_core==2.18.2
typing_extensions==4.11.0
requests==2.31.0
